* use VariablePair and VariableContainer in InferenceBatchLearners  * ParallelInferenceBatchLearner (Untested yet)  * New base class: ParallelVoteInference  * Remove InferenceBatchLearner file

+++ trunk/src/Class/TupleClass.h
@@ -49,9 +49,12 @@ public:
   virtual void destroy(VariableValue& value) const
   {
     Variable* data = (Variable* )value.getRawData();
-    for (size_t i = 0; i < size; ++i)
-      data[i] = Variable();
-    RawDataBuiltinTypeClass::destroy(value);
+    if (data)
+    {
+      for (size_t i = 0; i < size; ++i)
+        data[i] = Variable();
+      RawDataBuiltinTypeClass::destroy(value);
+    }
   }
 
   virtual String toString(const VariableValue& value) const
